job_agent/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests
beautifulsoup4
pydantic